move: conversation items to own file
