+++ app/src/main/java/me/danlowe/meshcommunicator/ui/screen/permissions/PermissionsScreen.kt
@@ -41,7 +41,7 @@ fun PermissionsScreen(
     )
 
     if (permissions.allPermissionsGranted) {
-        LaunchedEffect(Unit) {
+        LaunchedEffect(permissions.allPermissionsGranted) {
             navHandler(PermissionsNavEvent.PermissionsGranted)
         }
     } else {
